fix(emit): hoist es5 computed class temps
